Fix POR compile issues
