+++ .changeset/clean-lies-repeat.md
@@ -1,0 +1,5 @@
+---
+'@chainlink/proof-of-reserves-adapter': patch
+---
+
+Fixes compile issues
